+++ Magic/src/main/java/com/elmakers/mine/bukkit/economy/VaultCurrency.java
@@ -5,13 +5,16 @@ import org.bukkit.configuration.ConfigurationSection;
 import com.elmakers.mine.bukkit.api.magic.CasterProperties;
 import com.elmakers.mine.bukkit.api.magic.Mage;
 import com.elmakers.mine.bukkit.api.magic.MageController;
+import com.elmakers.mine.bukkit.api.magic.Messages;
 import com.elmakers.mine.bukkit.integration.VaultController;
 
 public class VaultCurrency extends BaseMagicCurrency {
     private final MageController controller;
+    private final boolean customFormatting;
 
     public VaultCurrency(MageController controller, ConfigurationSection configuration) {
         super(controller, "currency", configuration);
+        customFormatting = configuration.getBoolean("custom_formatting");
         this.controller = controller;
     }
 
@@ -57,5 +60,37 @@ public class VaultCurrency extends BaseMagicCurrency {
     @Override
     protected boolean hasDecimals() {
         return true;
+    }
+
+    @Override
+    public String getName(Messages messages) {
+        if (customFormatting) {
+            return super.getName(messages);
+        }
+        return VaultController.getInstance().getCurrencyPlural();
+    }
+
+    @Override
+    public String getShortName(Messages messages) {
+        if (customFormatting) {
+            return super.getShortName(messages);
+        }
+        return VaultController.getInstance().getCurrency();
+    }
+
+    @Override
+    public String getSingularName(Messages messages) {
+        if (customFormatting) {
+            return super.getSingularName(messages);
+        }
+        return VaultController.getInstance().getCurrency();
+    }
+
+    @Override
+    public String formatAmount(double amount, Messages messages) {
+        if (customFormatting) {
+            return super.formatAmount(amount, messages);
+        }
+        return VaultController.getInstance().format(amount);
     }
 }